Add renderErrorsAsMap to Components
In a Composer, it is now possible to do:
self.renderErrorsAsMap(bean:map)
This does the same as renderErrors, but instead of taking domain class,
it takes a list of maps. The map is structured as follows:
[field: field, error: error].
This is useful whenever one is using an api that returns json , or a
map.